LayoutTests/fast/encoding/char-encoding-expected.txt

This tests encoding characters in various character sets.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS encode('UTF-8', 'U+00A0') is '%C2%A0'
PASS encode('GBK', 'U+00A5') is '%26%23165%3B'
PASS encode('gb2312', 'U+00A5') is '%26%23165%3B'
PASS encode('GB_2312-80', 'U+00A5') is '%26%23165%3B'
PASS encode('EUC-CN', 'U+00A5') is '%26%23165%3B'
PASS encode('GBK', 'U+20AC') is '%80'
PASS encode('gb2312', 'U+20AC') is '%80'
PASS encode('GB_2312-80', 'U+20AC') is '%80'
PASS encode('EUC-CN', 'U+20AC') is '%80'
PASS encode('GBK', 'U+01F9') is '%A8%BF'
PASS encode('GBK', 'U+1E3F') is '%A8%BC'
PASS encode('GBK', 'U+22EF') is '%26%238943%3B'
PASS encode('GBK', 'U+301C') is '%26%2312316%3B'
PASS encode('csiso2022kr', 'U+00A0') is '%C2%A0'
PASS encode('hz-gb-2312', 'U+00A0') is '%C2%A0'
PASS encode('iso-2022-cn', 'U+00A0') is '%C2%A0'
PASS encode('iso-2022-cn-ext', 'U+00A0') is '%C2%A0'
PASS encode('iso-2022-kr', 'U+00A0') is '%C2%A0'
PASS successfullyParsed is true

TEST COMPLETE
